requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
plotly>=5.3.0
scipy>=1.7.0
pywavelets>=1.1.0
scikit-learn>=1.0.0
torch>=1.9.0
opencv-python>=4.5.0
streamlit>=1.10.0
pillow>=8.0.0